Moved WADL tests to resteasy-wadl module and fixed/refactored them to remove the @Ignore annotations
