+++ resteasy-wadl/src/test/java/org/jboss/resteasy/test/nextgen/wadl/resources/BasicResource.java
