+++ resteasy-wadl/src/test/java/org/jboss/resteasy/test/nextgen/wadl/resources/issues/RESTEASY1246.java
